move data/constants to lib/constants
