+++ src/pages/index.tsx
@@ -11,7 +11,7 @@ import get from 'utils/get'
 import Header from 'components/Header'
 import HeadingSection from 'components/HeadingSection'
 import MoviesList from 'components/MoviesList/index'
-import { LANGUAGES } from 'data/constants'
+import { LANGUAGES } from '~/lib/constants'
 import { LanguageProvider } from 'context/LanguageContext.ts'
 import useLocalStorage from 'hooks/useLocalStorage'
 import { createMeilisearchClient } from '../lib/createMeilisearchClient'
