+++ src/stories/Radio.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Radio from 'components/Radio'
 import { useRadioState, RadioGroup } from 'reakit/Radio'
-import { LANGUAGES } from 'data/constants'
+import { LANGUAGES } from '~/lib/constants'
 
 export default {
   title: 'Components/Radio',
